+++ yukon-server/CapControl/include/clistener.h
@@ -25,7 +25,7 @@
 class CtiCCClientListener : public CtiObservable, public CtiObserver
 {
 public:
-    CtiCCClientListener(UINT port);
+    CtiCCClientListener(LONG port);
     virtual ~CtiCCClientListener();
 
     virtual void start();
@@ -39,7 +39,7 @@ protected:
 private:
     RWSocketListener* _listener;
 
-    UINT _port;   
+    LONG _port;   
     RWThread _listenerthr;
     RWThread _checkthr;
