+++ yukon-server/LoadManagement/include/clistener.h
@@ -25,7 +25,7 @@
 class CtiLMClientListener
 {
 public:
-    CtiLMClientListener(UINT port);
+    CtiLMClientListener(LONG port);
     virtual ~CtiLMClientListener();
 
     virtual void start();
@@ -40,7 +40,7 @@ protected:
 private:
     RWSocketListener* _socketListener;
 
-    UINT _port;   
+    LONG _port;   
     RWThread _listenerthr;
     RWThread _checkthr;
